38. Opravena přístupová práva na editaci modelu v modelu User 	- Všechny kopie a D&D záznamy Typentity a Attribute označeny "system=false" po akci přidání do DB 37. Opraveny testy na práva editace, přidávání a mazání Typentity, Attribute     - založen objekt User (ten je nutné @Injekt do dalších Beans, aby se dal používat)     - Nakonfigurovat dialog pro přidávání sestav k Typentity 36. Opraveno ukládání ModelDetail.Attribute >> přes ModelTable     -  ModelTree vylepšit validaci na jedinečnost Zkratky a platnosti datumů OD-DO - přidat "placeholder"     -  ModelDetail vylepšit validaci na platnosti datumů OD-DO - přidat "placeholder"
32. Chyba v překreslování detailu byla ve špatné definici update, místo @Form musí být :modelFormDetail
    - vyřešit obnovení hodnoty v detailu, pokud je chyba validace a uživatel si v tabulce zvolí jiný řádek s Attribute

30. Chybuje ukládání Attribute do DB ModelDetail, ošetřit validaci datumů PlatiOd-PlatiDo
    - mazání položky Attribute v ModelTable + přidávání Attribute

27. Úkoly zůstávají , započata práce na ModelTable (Attribute) plnění doplněno o načtení 1 nové Attribute
	- vyřešit zápis po editaci detailu a mazání položky Attribute z tabulky (tlačítko "Delete") dát k detailu
	- dopracovat v ModelTree vložení položky Typentity s polem "system" jako false, protože to voložil uživatel - zvážit dopad
	- zvážit přechod k uživatelsky definovaným transakcím
	- vyladit error hlášení
	- vyladit logování přechodem na Log4j
23. Rozpracováno paste TreeNode - nutno dodělat klonování Attribute, zapisování změn je zpožděno-vyřešit getEm.flush() zápisu - jak to funguje

21. Dodělat předchozí - D&D - vyřešit uložení do DB při kopii TreeNode (nové ID, IDparent, attribute)
    - copy a paste jak na zásobník viz předchozí
    - nefunguje "disable" v MenuItems kontextového menu - opraveno - nutné provest UpDate menu po každé Ajax události "Select" 
20. Odladit D&D, práce s DB - trasakce, Error message
19. Dodělat .html a ajax ukládáním dat, ukládání změn do DB, D&D
18. Opraveno přidávání pole, je divné chování vlastnosti valueChangeListener
17. Deatil přidává nová pole, jako by focus mělo pořád tlačítko Add
16. Detail typentity
15. Napojit všechno na DB a napsat editaci
14. Rozpracovat Attr detail
13. treeTable - dodělat DragAndDrop, CopyAndPaste
12. Rozpracovaný modelTable pro Attributy - dodělat DragAndDrop, Copy and Paste
10. Doma nastavena konfigurace WildFly, která funguje.
9. Dopracovat modelTree (HTML je nachystane), rozpracovat tabulku Attr a detail Attr
8. Po rekonfiguraci WildFly začalo fungovat Persistence Api s Eclipse

3. Hlavní menu - vyladit
2. Pokus o Hlavní menu
1. Upravy souboru .CSV a .XM

+++ src/main/java/cz/rental/admin/model/ModelTree.java
@@ -5,6 +5,7 @@
  */
 package cz.rental.admin.model;
 
+import cz.rental.aplikace.User;
 import cz.rental.entity.Typentity;
 import java.io.Serializable;
 import java.util.logging.Level;
@@ -63,7 +64,6 @@ public class ModelTree implements Serializable {
         // System.out.println(" onNodeSelect(NodeSelectEvent event): typentity.getTypentity(): " + typentity.getTypentity());
     }
 
-
     /**
      * Vymaze <CODE>selectedNode</CODE> z DB a z TreeNode + vymaz vsech potomku
      * a jejich Typentity (ulozeno v TreeNode.getData() ) a Attribute + vymaz
@@ -87,6 +87,9 @@ public class ModelTree implements Serializable {
     private void deleteFromDb(TreeNode selectedNode) {
         try {
             Typentity typEntityDel = (Typentity) selectedNode.getData();
+            if (typEntityDel.getAttrsystem() != null && typEntityDel.getAttrsystem() && !user.getParam(User.SUPERVISOR, false)) {
+                return;
+            }
             this.attrController.deleteAllTypentityId(typEntityDel.getId());
             this.controller.destroy(typEntityDel);
             for (TreeNode treeNode : selectedNode.getChildren()) {
@@ -183,7 +186,7 @@ public class ModelTree implements Serializable {
      * @return uzivatel muze pridavat zaznamy
      */
     public boolean isAppendable() {
-        boolean isAppendable = user.getParam("SUPERVISOR", false) || user.getParam("AddTypentity", false);
+        boolean isAppendable = user.getParam(User.SUPERVISOR, false) || user.getParam(cz.rental.aplikace.User.MODEL_EDIT, false);
         return (isAppendable);
     }
 
@@ -196,9 +199,9 @@ public class ModelTree implements Serializable {
     public boolean isEditable() {
         boolean isEditable = (this.typentity != null);
         if (isEditable) {
-            isEditable = user.getParam("SUPERVISOR", false);
+            isEditable = user.getParam(User.SUPERVISOR, false);
             if (!isEditable) {
-                isEditable = this.typentity.getAttrsystem() != null && !this.typentity.getAttrsystem() && user.getParam("EditTypentity", false);
+                isEditable = this.typentity.getAttrsystem() != null && !this.typentity.getAttrsystem() && user.getParam(cz.rental.aplikace.User.MODEL_EDIT, false);
             }
         }
         return (isEditable);
@@ -212,9 +215,9 @@ public class ModelTree implements Serializable {
     public boolean isRemovable() {
         boolean isRemoveable = (this.typentity != null);
         if (isRemoveable) {
-            isRemoveable = user.getParam("SUPERVISOR", false);
+            isRemoveable = user.getParam(User.SUPERVISOR, false);
             if (!isRemoveable) {
-                isRemoveable = this.typentity.getAttrsystem() != null && !this.typentity.getAttrsystem() && user.getParam("DelTypentity", false);
+                isRemoveable = this.typentity.getAttrsystem() != null && !this.typentity.getAttrsystem() && user.getParam(cz.rental.aplikace.User.MODEL_EDIT, false);
             }
         }
         return (isRemoveable);
@@ -287,6 +290,7 @@ public class ModelTree implements Serializable {
     private TreeNode pasteNode(TreeNode node, TreeNode nodeParent, int parentIndex) throws Exception {
         Typentity typEntityNew = controller.cloneTypentity((Typentity) node.getData());
         typEntityNew.setIdparent(((Typentity) nodeParent.getData()).getId());
+        typEntityNew.setAttrsystem(false);
         controller.create(typEntityNew);
         typEntityNew.setNewEntity(false);
         attrController.cloneAttribute((Typentity) node.getData(), typEntityNew);
@@ -381,5 +385,4 @@ public class ModelTree implements Serializable {
     public void setCopyNode(TreeNode copyNode) {
         this.copyNode = copyNode;
     }
-
 }
